Refactored to avoid circular references when checking github events
